requirements.txt
fastapi
uvicorn
python-dotenv
aiohttp
pydantic
google-cloud-aiplatform
vertexai
loguru
